extract: dont create new directories
Extract would create directories as if it were going to put the
extracted files in there. And then it didnt. How did i miss that?

In the future, extract should have an option to create these dirs and
then put the extracted files in there. Just commented out for now.